play.py 파일 이름 수정, 로비에서 시작 버튼 누르면 게임 실행되게 설정
